+++ src/main/java/com/brano/ecomonitor/controller/PollutantController.java
@@ -1,15 +1,10 @@
 package com.brano.ecomonitor.controller;
 
-import com.brano.ecomonitor.entity.Company;
-import com.brano.ecomonitor.entity.Pollutant;
-import com.brano.ecomonitor.model.PollutantModel;
+import com.brano.ecomonitor.dto.PollutantModel;
 import com.brano.ecomonitor.service.PollutantService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.ModelAndView;
-
-import java.util.List;
 
 import static com.brano.ecomonitor.controller.ResponseWrapper.wrap;
 
@@ -40,7 +35,7 @@ public class PollutantController {
     @DeleteMapping("/pollutant")
     public ResponseEntity<?> deletePollutant(@RequestParam String id) {
         return wrap(() -> {
-            pollutantService.deleteById(Long.parseLong(id));
+            pollutantService.deleteById(Integer.parseInt(id));
             return "Success";
         }, HttpStatus.ACCEPTED);
     }
