+++ src/main/java/com/brano/ecomonitor/repository/PollutionRepository.java
@@ -8,17 +8,17 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PollutionRepository extends JpaRepository<Pollution, Long> {
+public interface PollutionRepository extends JpaRepository<Pollution, Integer> {
 
     @Query("SELECT DISTINCT p.year FROM Pollution p")
-    List<Long> findAllDistinctYear();
+    List<Integer> findAllDistinctYear();
 
-    List<Pollution> findAllByCompanyCompanyId(Long id);
+    List<Pollution> findAllByCompanyCompanyId(Integer id);
 
-    List<Pollution> findAllByCompanyCompanyIdAndYear(Long id, Integer year);
+    List<Pollution> findAllByCompanyCompanyIdAndYear(Integer id, Integer year);
 
-    void deleteAllByCompanyCompanyId(Long id);
+    void deleteAllByCompanyCompanyId(Integer id);
 
-    void deleteAllByPollutantPollutantId(Long id);
+    void deleteAllByPollutantPollutantId(Integer id);
 
 }
